GDAPS II Final Project
This group project at the end of my first year was created in Visual Studio using C# in a group of 4 persons. I handled game balance, ability creation, and ability implementation in relation to what users see and experience (adjusting health and stamina when users activate attacks). The project was tremendous fun and is the spiritual successor to a passion project I made in high school as both games use similar themes and mechanics.