RPG Seminar Project
Working in a group of four, friends and I created a tabletop roleplaying game from scratch. The challenges we attempted to address in our project were to not use d20s (twenty-sided die) and address the issue of having stagnant characters in a multi-session long campaign. Our solution to the former was to simply use other dice while the latter was addressed by playing not as a single character but rather that character and their descendants. We spent a semester of school iterating on deigns and working through our project to yield this final product. The entire rules for the game are published and free to download at the below link:

https://www.drivethrurpg.com/product/261278/Bloodline