The Dread Surgeon
This semester-long project between myself and a friend had the goal of creating scenarios that could be easily adapted for differently leveled players (of D&D 5e) to enjoy. The project involves custom NPCs, Items, and scenarios to produce material suited for a 1-shot roleplaying encounter. Encounters are balanced to be held at 1st, 9th, or 13+th level so that a group of players can enjoy regardless of their characters' experience. This tackled the personal problem as a GM of needing to manually tune encounters if the party was too powerful or weak when I wanted to run a specific, fun encounter. The result was this adventure.