Custom Hearthstone Expansion (WIP)
As a passion project, I designed and have been iterating on a custom Hearthstone expansion featuring one of my favorite themes in Warcraft: The Old Gods. To this end, I have created my own Old Gods and developed an expansion around them including the Return of Madame Lazul, the might of the Scarlet Crusade, and the Encounter keyword where players summon minions for their opponent so that they can defeat them. These 'Encounter' minions play into the theme of accepting the challenge of a new foe. Encounter minions, however, will benefit the player that played them as they tear apart their controller or otherwise benefit the card's player.

This set is undergoing many iterations as I get feedback from fellow Hearthstone enthusiasts and Game Designers. My end-goal with this project is to make something that I'm proud of and that I can use to show my design process as I keep track of previous iterations.

The following is a view-only link to the spreadsheet:
https://docs.google.com/spreadsheets/d/1aTM4Fb_qn2mJgz1boefhsXlo3TMxlWzrt76-iQ4TcdE/edit?usp=sharing